added pagination in hearing home Screen and clean code in Home hearing tab page (#3587)
