+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/home/src/pages/employee/MainHomeScreen.js
@@ -335,6 +335,7 @@ const MainHomeScreen = () => {
               setFilters={setFilters}
               filters={filters}
               showToast={showToast}
+              hearingCount={hearingCount}
             />
           </div>
         ) : (
